backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
supabase
langgraph
langchain-openai
langchain-core
httpx
